+++ tests/Unit/Segments/LINLineItemTest.php
@@ -21,4 +21,25 @@ class LINLineItemTest extends TestCase
         self::assertEquals('2', $segment->subId());
         self::assertEquals($rawValues, $segment->rawValues());
     }
+
+    /**
+     * @test
+     */
+    public function typed_accessors(): void
+    {
+        $segment = new LINLineItem(['LIN', '1', '', ['5449000000996', 'EN']]);
+
+        self::assertSame('1', $segment->lineNumber());
+        self::assertSame(['5449000000996', 'EN'], $segment->itemNumberIdentification());
+        self::assertSame('5449000000996', $segment->itemNumber());
+        self::assertSame('EN', $segment->itemTypeCode());
+    }
+
+    /**
+     * @test
+     */
+    public function item_number_identification_is_empty_when_absent(): void
+    {
+        self::assertSame([], (new LINLineItem(['LIN', '1']))->itemNumberIdentification());
+    }
 }